Correct the functionality of the test app

+++ src/testApplication.h
@@ -18,8 +18,8 @@ struct taskProperties
     int taskPriority;
     int taskWorstCaseExecuteTime;
     int xDelay;
-    long xFibonnaciCycles;
-    long xFibonnaciCyclesWorstCase;
+    unsigned int xFibonnaciCycles;
+    unsigned int xFibonnaciCyclesWorstCase;
     int *xPowerConsumptionTestIsWorstCase;
 };
 #ifdef __cplusplus
